a function that iterate an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include "function_pointers.h"
 
 /**
@@ -10,7 +11,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
 
 	if (array && action)
 	{
